+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v0.2.0 (in development)
 - Sanitize all ASCII non-printable characters in torrent names when filling in
   output path templates
 - Suppress sub-INFO log messages from dependencies
+- Fix a Heisenbug involving attrs, slotted classes, and garbage collection
 
 v0.1.0 (2023-06-11)
 -------------------